requirements.txt
allure-pytest
allure-python-commons
pycodestyle
pytest-xdist
requests
selenium
